backend/requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
fastapi==0.104.1
feedparser==6.0.10
newspaper3k==0.2.8
pydantic==2.4.2
uvicorn==0.24.0
python-multipart==0.0.6
python-jose==3.3.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0